feat(remove-editor): remove unused mocks
- Delete outdated mocks (projects, integrations, routers, and types).
- Remove unused GitLab repository integrations.
- Clean up unnecessary project configurations and history data.
