--- packages/portal/server/mocks/integrations/integrations.ts
@@ -1,48 +1,0 @@
-/**
- * Copyright 2024-2025 NetCracker Technology Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { IntegrationBranchesDto, IntegrationRepositoriesDto } from './types'
-
-export const REPOSITORIES: IntegrationRepositoriesDto = {
-  repositories: [
-    {
-      repositoryId: 'repository-1',
-      name: 'APIHUB/apihub-ui',
-      defaultBranch: 'main',
-    },
-    {
-      repositoryId: 'repository-2',
-      name: 'APIHUB/api-hub-backend',
-      defaultBranch: 'main',
-    },
-    {
-      repositoryId: 'repository-3',
-      name: 'APIHUB/apihub-configuration',
-      defaultBranch: 'main',
-    },
-  ],
-}
-
-export const BRANCHES: IntegrationBranchesDto = {
-  branches: [
-    {
-      name: 'main',
-    },
-    {
-      name: 'develop',
-    },
-  ],
-}
