--- packages/portal/server/mocks/integrations/types.ts
@@ -1,33 +1,0 @@
-/**
- * Copyright 2024-2025 NetCracker Technology Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export type IntegrationRepositoryDto = Readonly<{
-  repositoryId: string
-  name: string
-  defaultBranch: string
-}>
-
-export type IntegrationRepositoriesDto = Readonly<{
-  repositories: ReadonlyArray<IntegrationRepositoryDto>
-}>
-
-export type IntegrationBranchDto = Readonly<{
-  name: string
-}>
-
-export type IntegrationBranchesDto = Readonly<{
-  branches: ReadonlyArray<IntegrationBranchDto>
-}>
