--- packages/portal/server/routers/integrations/gitlab-repositories.ts
@@ -1,30 +1,0 @@
-/**
- * Copyright 2024-2025 NetCracker Technology Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { BRANCHES, REPOSITORIES } from '../../mocks/integrations/integrations'
-import type { IntegrationsRouter } from './router'
-
-export function getGitlabRepositories(router: IntegrationsRouter): void {
-  router.get('/gitlab/repositories/', (req, res) => {
-    res.status(200).json(REPOSITORIES)
-  })
-}
-
-export function getGitlabRepositoryBranches(router: IntegrationsRouter): void {
-  router.get('/gitlab/repositories/:id/branches/', (req, res) => {
-    res.status(200).json(BRANCHES)
-  })
-}
